Polyline interpolation animation, track visibility

+++ MovieMorpher/MorphFBOprocessor.cpp
@@ -87,8 +87,13 @@ void MorphFBOprocessor::Render()
 			glUniform1fARB(glsl_pipeline.GPUPrograms["morph"]->GetUniform("fMorphPower"),  fMorphPower);
 			glUniform1fARB(glsl_pipeline.GPUPrograms["morph"]->GetUniform("fMorphRatio"),  fMorphRatio);
 
-					RenderTexturedQuadMesh(texBank[TEXTURE_INPUT_IMAGE]->m_uiTextureID,
-										   -m_iWidth/2, -m_iHeight/2, m_iWidth, m_iHeight, 0.0f, 110*float(m_iWidth)/float(m_iHeight), 110 );
+					RenderTexturedQuadMesh( texBank[TEXTURE_INPUT_IMAGE]->m_uiTextureID,
+										    float(-m_iWidth/2),
+											float(-m_iHeight/2),
+											float(m_iWidth),
+											float(m_iHeight),
+											0.0f,
+											int(110*float(m_iWidth)/float(m_iHeight)), 110 );
 
 			// if (bCPUShaderDebugging) {
 			//	   GenMeshToList(-m_iWidth / 2.0f, -m_iHeight / 2.0f, m_iWidth, m_iHeight, 0.0, 30, 30);
